refactor(graphics): split graphics helpers by config vs launch, refactor code structure
