+++ src-tauri/src/launch/helpers/mod.rs
@@ -1,5 +1,6 @@
 pub mod command_generator;
 pub mod file_validator;
+pub mod graphics_handler;
 pub mod jre_selector;
 pub mod log_parser;
 pub mod misc;
