+++ src-tauri/src/launcher_config/helpers/mod.rs
@@ -1,3 +1,4 @@
+pub mod graphics;
 pub mod java;
 pub mod misc;
 pub mod updater;
