+++ src-tauri/src/lib.rs
@@ -87,6 +87,7 @@ pub async fn run() {
         launcher_config::commands::retrieve_java_list,
         launcher_config::commands::validate_java,
         launcher_config::commands::download_mojang_java,
+        launcher_config::commands::retrieve_supported_graphics_renderers,
         launcher_config::commands::check_game_directory,
         launcher_config::commands::clear_download_cache,
         launcher_config::commands::check_launcher_update,
@@ -149,7 +150,6 @@ pub async fn run() {
         launch::commands::validate_game_files,
         launch::commands::validate_selected_player,
         launch::commands::launch_game,
-        launch::commands::retrieve_supported_graphics_renderers,
         launch::commands::cancel_launch_process,
         launch::commands::open_game_log_window,
         launch::commands::retrieve_game_log,
