+++ src/locales/en.json
@@ -903,67 +903,6 @@
         }
       }
     },
-    "graphics": {
-      "title": "Graphics",
-      "settings": {
-        "api": {
-          "title": "Graphics API",
-          "default": {
-            "label": "Default",
-            "desc": "Use Minecraft and the driver default."
-          },
-          "opengl": {
-            "label": "OpenGL",
-            "desc": "Use OpenGL renderer and optional Mesa driver overrides."
-          },
-          "vulkan": {
-            "label": "Vulkan",
-            "desc": "Use Vulkan renderer and optional ICD driver selection."
-          }
-        },
-        "renderer": {
-          "title": "Renderer / Driver",
-          "default": {
-            "label": "Default",
-            "desc": "Do not set renderer-specific environment variables."
-          },
-          "llvmpipe": {
-            "label": "LLVMpipe",
-            "desc": "Mesa software OpenGL renderer."
-          },
-          "zink": {
-            "label": "Zink",
-            "desc": "Run OpenGL on top of Vulkan through Mesa."
-          },
-          "d3d12": {
-            "label": "D3D12",
-            "desc": "Mesa OpenGL driver based on DirectX 12."
-          },
-          "lavapipe": {
-            "label": "Lavapipe",
-            "desc": "Mesa software Vulkan driver."
-          },
-          "dozen": {
-            "label": "Dozen",
-            "desc": "Mesa Vulkan driver based on DirectX 12."
-          },
-          "nvidia_vulkan": { "label": "NVIDIA Vulkan" },
-          "nvidia_nvk": { "label": "NVK" },
-          "amdvlk": { "label": "AMDVLK" },
-          "amd_radv": { "label": "RADV" },
-          "intel_vulkan": { "label": "Intel Vulkan" },
-          "intel_anv": { "label": "ANV" },
-          "intel_hasvk": { "label": "HASVK" },
-          "qualcomm": { "label": "Qualcomm Vulkan" },
-          "turnip": { "label": "Turnip" },
-          "moltenvk": { "label": "MoltenVK" },
-          "kosmickrisp": { "label": "KosmicKrisp" },
-          "powervr": { "label": "PowerVR" },
-          "panvk": { "label": "PanVK" },
-          "v3dv": { "label": "V3DV" }
-        }
-      }
-    },
     "jvm": {
       "title": "Java VM Options",
       "settings": {
@@ -988,6 +927,49 @@
         },
         "args": {
           "title": "Other JVM Arguments"
+        }
+      }
+    },
+    "graphics": {
+      "title": "Graphics",
+      "settings": {
+        "api": {
+          "title": "Graphics API",
+          "default": {
+            "label": "Default",
+            "desc": "Use Minecraft and the driver default."
+          },
+          "opengl": {
+            "label": "OpenGL",
+            "desc": "Use OpenGL rendering."
+          },
+          "vulkan": {
+            "label": "Vulkan",
+            "desc": "Use Vulkan rendering."
+          }
+        },
+        "renderer": {
+          "title": "Renderer / Driver",
+          "default": "Default",
+          "llvmpipe": "LLVMpipe",
+          "zink": "Zink",
+          "d3d12": "D3D12",
+          "lavapipe": "Lavapipe",
+          "dozen": "Dozen",
+          "nvidia_vulkan": "NVIDIA Vulkan",
+          "nvidia_nvk": "NVK",
+          "amdvlk": "AMDVLK",
+          "amd_radv": "RADV",
+          "intel_vulkan": "Intel Vulkan",
+          "intel_anv": "ANV",
+          "intel_hasvk": "HASVK",
+          "qualcomm": "Qualcomm Vulkan",
+          "turnip": "Turnip",
+          "moltenvk": "MoltenVK",
+          "kosmickrisp": "KosmicKrisp",
+          "powervr": "PowerVR",
+          "panvk": "PanVK",
+          "v3dv": "V3DV"
         }
       }
     },
